Fixed entity attack to int

+++ src/Entity.java
@@ -20,7 +20,7 @@ public class Entity extends GraphicsProgram {
 
 	private int level;
 	private int HP; // health points
-	public double attack;
+	public int attack;
 	private int exp; //experience
 	private String name;
 	private EntityType type;
@@ -62,11 +62,11 @@ public class Entity extends GraphicsProgram {
 	public void death() {
 		//TODO
 	}
-	public double getAttack() {
+	public int getAttack() {
 		
 		return attack;
 	}
-	public void setAttack(double atk) {
+	public void setAttack(int atk) {
 		
 		attack = atk;
 	}
